URLs with reference (#) are now checked.
There is a limitation though. Imagine that we want to check "http://host/index.html#ref1". I need to check if in index.html there is something like this: <a name="ref1"></a> but if index.html has not been checked yet I cannot know it, for example if the root URL is http://host/dummy.html which has in his html document: "<a href="index.html#ref1></a>". When that last tag is checked only dummy.html have been checked so I cannot know.

It's late and I don't know if my explanation makes any sense, but I think it does ;)

svn path=/trunk/kdewebdev/klinkstatus/; revision=313116