The most non trivial port to Qt 4 is done: K3ListView -> QTreeWidget. Not much left I think. Lots of fixes and refactorings in the way, that should had been done in trunk (if I didn't have a job :)).
HTML validation initial work is done, i.e., there is a (configurable) column which indicates with an icon whether the document has fixable or not fixable HTML errors. Tooltips should be implemented so the user can find out what the icon means (the same for other columns).
The validation engine uses HTML Tidy. Still have to test if the cmake check works for windows.

Next step will be adding the possibility of correcting all HTML documents in a row, and allowing to fix a document individually via Kompare part.

svn path=/branches/work/klinkstatus4-tidy/; revision=641991