Advance a little more on unreferred documents and HTML validation.
Set the UI infrastructure to support the new features. Now the session widget is a stack widget that necessarily contains a SessionWidget (the name is not clear now has each session can be constituted by several widgets) object. The stack widget can also contain a widget for unreferred documents, for HTML validation and so on. The stack widget is now the page for each tab.
Some care have been take to deal with the actions. Not very hard though as ActionManager was already well designed.

A consequence of this change is the refactoring of TabWidgetSession which was a little stupid btw.

svn path=/trunk/KDE/kdewebdev/klinkstatus/; revision=655210