First steps for the automation dialog GUI:
Use KConfigXT instead of plain KConfig. Love the kcfgfile arg="true" feature in the kcfg file which allows to construct multiple instances pointing to different files, which is needed here because each site check to automate has its own configuration file. The implementation is a little strange though :( I have to do things like this:

    AutomationConfig::instance(QString());
    delete AutomationConfig::self();
    AutomationConfig::instance(configurationFile);
    AutomationConfig* config = AutomationConfig::self();

because AutomationConfig::instance only do something if self() is a null pointer.

The automation configuration dialog now shows with the several site entries in the left, still missing the widget with the configurations and the wizard to configure a new automated site check.

CCMAIL: Sam Ryan <thesquib@gmail.com>

svn path=/trunk/KDE/kdewebdev/klinkstatus/; revision=792290